Remove optionbuilder, extensions, and customizer

+++ tests/ConsoleSample/Migrations/20240503230205_Initial.cs
@@ -27,7 +27,7 @@ namespace ConsoleSample.Migrations
                 name: "CustomFields",
                 columns: table => new
                 {
-                    Id = table.Column<int>(type: "int", nullable: false)
+                    Id = table.Column<long>(type: "bigint", nullable: false)
                         .Annotation("SqlServer:Identity", "1, 1"),
                     Name = table.Column<string>(type: "nvarchar(max)", nullable: false),
                     CustomFieldType = table.Column<int>(type: "int", nullable: false),
