+++ tests/ConsoleSample/Migrations/AppDbContextModelSnapshot.cs
@@ -38,11 +38,11 @@ namespace ConsoleSample.Migrations
 
             modelBuilder.Entity("EFCoreCustomFields.CustomField", b =>
                 {
-                    b.Property<int>("Id")
+                    b.Property<long>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("int");
+                        .HasColumnType("bigint");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"));
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("Id"));
 
                     b.Property<int>("CustomFieldType")
                         .HasColumnType("int");
